Reorganise code in Processor into other classes
Add Parser.java and TaskManager.java to handle command parsing and editing the task list
