--- data/db.txt
@@ -1,2 +1,0 @@
-D||0||0||01-01-2003
-D||0||0||20-10-2001
