+++ src/main/java/Angela.java
@@ -5,7 +5,7 @@ public class Angela {
     public static void main(String[] args) {
         Storage storage = new Storage();
         TaskManager manager = new TaskManager();
-        Parser parser = new Parser(storage);
+        Parser parser = new Parser();
         Processor process = new Processor(storage, manager, parser);
         Output.introOutput();
         process.start();
